Flytt ut openapi, ikke muter mapper
